move api call and data calculations to main
